pyproject.toml
[tool.poetry]
name = "instaclone-django"
version = "0.1.0"
description = ""
authors = ["wsxqaz789456 <sindaea77@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "instaclone_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
colorama = "^0.4.6"
boto3 = "^1.26.60"
django-cors-headers = "^3.13.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
